Add websocket support, allowing TeamContext to update in real-time
Whenever certain changes are made to the database, we use postgres
NOTIFY to send an event to the node app containing a { table, op, row }
payload. The backend app uses a separate pg Client to listen for these
events and dispatch events to connected websocket clients that have
subscribed to the relevant team (by providing the team code in a
websocket request). In the frontend, TeamContext initializes a WebSocket
client and updates its own state in response to these events, triggering
a re-render of the updated components.

In order for nginx to support websocket requests, they need to be routed
to the backend, and explicit support must be enabled to upgrade to
HTTP 1.1.