Implement websocket ping/pong heartbeat to keep connections alive
The client (testing in Chrome 80.0.3987.149) will disconnect after 60
seconds of inactivity (or potentially this is nginx dropping the
connection; not sure). With this change, our backend app sends a ping
to each websocket client every 5 seconds: if we don't get a pong
response by the next check, we drop the client. This hearbeat mechanism
keeps the connection from being dropped due to inactivity.