Simplify SQL queries used in team API
The database operations in our API routes naturally involve some
indirection based on the fact that users supply a team code for basic
authentication:

- For all team-related routes, we need to look up the relevant team
  from its string code, and abort immediately if there's no such team

- When updating player state (including piece_ids in the fossil.have
  association table), we need to validate that request contains the
  right team code for the given player ID

- When querying team data, we need to do multiple lookups:
  - Get team row (including team id) from team code
  - With that team id: Get player rows (including player id)
  - With those player ids: Get piece ids for each player

Previously, we attempted to cram all this validation and these multiple
lookups into a single statement for each route. This turned out to be a
misguided way of implicitly building atomicity into our routes. If we
need our API operations to be atomic, we can just use actual SQL
transactions.

This change breaks several of our overly-complex SQL queries into
multiple steps. Rather than wrapping a kitchen-sink query into a single
middleware function within the route, we instead use the new
validateTeamCode or validatePlayerId functions to handle prerequisite
validation, and return a 404 if unable to proceed.

Before-and-after profiling results for GET /api/team/:code are here:
https://pastebin.com/raw/gCR3wa8G

In short, a GET request for a single team's data used to look like this:

    WITH t AS (SELECT * FROM fossil.team WHERE code = '833HZFubQ')
    SELECT
      (SELECT id FROM t) AS team_id,
      (SELECT name FROM t) AS team_name,
      (SELECT code FROM t) AS team_code,
      player.id AS player_id,
      player.name AS player_name,
      piece_id
    FROM fossil.player LEFT JOIN fossil.have
      ON have.player_id = player.id
    WHERE team_id = (SELECT id FROM t) ORDER BY player.id;

Whereas we now run three separate queries:

    SELECT * FROM fossil.team WHERE code = '833HZFubQ';
    SELECT id, name FROM fossil.player WHERE team_id = 3;
    SELECT player_id, json_agg(piece_id) AS pieces FROM fossil.have
      WHERE player_id IN (4, 5, 3, 6, 7, 8, 9, 10, 11, 12)
      GROUP BY player_id;

The result is slightly faster to execute in PostgreSQL, but the more
significant factor is that it returns far less duplicated data, and the
results are formatted much more closely to how the API represents them,
so there's a lot less mess in the backend app.